+++ example/jls/jls_util_prv.h
@@ -55,4 +55,5 @@ int on_help(struct app_s * self, int argc, char * argv[]);
 int on_fsr_statistics(struct app_s * self, int argc, char * argv[]);
 int on_info(struct app_s * self, int argc, char * argv[]);
 int on_inspect(struct app_s * self, int argc, char * argv[]);
+int on_read_fuzzer(struct app_s * self, int argc, char * argv[]);
 int on_version(struct app_s * self, int argc, char * argv[]);